+++ src/main/java/org/owasp/webgoat/lessons/missingac/MissingFunctionACHiddenMenus.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.missing_ac;
+package org.owasp.webgoat.lessons.missingac;
 
 import org.owasp.webgoat.container.assignments.AssignmentEndpoint;
 import org.owasp.webgoat.container.assignments.AssignmentHints;
@@ -29,32 +29,33 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import static org.owasp.webgoat.lessons.missing_ac.MissingFunctionAC.PASSWORD_SALT_ADMIN;
-
+/**
+ * Created by jason on 1/5/17.
+ */
 @RestController
-@AssignmentHints({"access-control.hash.hint6", "access-control.hash.hint7",
-        "access-control.hash.hint8", "access-control.hash.hint9", "access-control.hash.hint10", "access-control.hash.hint11", "access-control.hash.hint12"})
-public class MissingFunctionACYourHashAdmin extends AssignmentEndpoint {
+@AssignmentHints({"access-control.hidden-menus.hint1","access-control.hidden-menus.hint2","access-control.hidden-menus.hint3"})
+public class MissingFunctionACHiddenMenus extends AssignmentEndpoint {
 
-    private final MissingAccessControlUserRepository userRepository;
-
-    public MissingFunctionACYourHashAdmin(MissingAccessControlUserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @PostMapping(path = "/access-control/user-hash-fix", produces = {"application/json"})
+    @PostMapping(path = "/access-control/hidden-menu", produces = {"application/json"})
     @ResponseBody
-    public AttackResult admin(String userHash) {
-        //current user should be in the DB
-        //if not admin then return 403
-
-        var user = userRepository.findByUsername("Jerry");
-        var displayUser = new DisplayUser(user, PASSWORD_SALT_ADMIN);
-        if (userHash.equals(displayUser.getUserHash())) {
-            return success(this).feedback("access-control.hash.success").build();
-        } else {
-            return failed(this).feedback("access-control.hash.close").build();
+    public AttackResult completed(String hiddenMenu1, String hiddenMenu2) {
+        if (hiddenMenu1.equals("Users") && hiddenMenu2.equals("Config")) {
+            return success(this)
+                    .output("")
+                    .feedback("access-control.hidden-menus.success")
+                    .build();
         }
-    }
 
+        if (hiddenMenu1.equals("Config") && hiddenMenu2.equals("Users")) {
+            return failed(this)
+                    .output("")
+                    .feedback("access-control.hidden-menus.close")
+                    .build();
+        }
+
+        return failed(this)
+                .feedback("access-control.hidden-menus.failure")
+                .output("")
+                .build();
+    }
 }
