+++ src/main/resources/lessons/missingac/html/MissingFunctionAC.html
@@ -1,12 +1,12 @@
 <html xmlns:th="http://www.thymeleaf.org">
 
 <div class="lesson-page-wrapper">
-    <div class="adoc-content" th:replace="doc:lessons/missing_ac/documentation/missing-function-ac-01-intro.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/missingac/documentation/missing-function-ac-01-intro.adoc"></div>
 </div>
 
 <div class="lesson-page-wrapper">
     <link rel="stylesheet" type="text/css" th:href="@{/lesson_css/ac.css}"/>
-    <div class="adoc-content" th:replace="doc:lessons/missing_ac/documentation/missing-function-ac-02-client-controls.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/missingac/documentation/missing-function-ac-02-client-controls.adoc"></div>
 
     <div class="attack-container">
         <nav class="navbar navbar-default">
@@ -70,7 +70,7 @@
 
 <div class="lesson-page-wrapper">
 
-    <div class="adoc-content" th:replace="doc:lessons/missing_ac/documentation/missing-function-ac-03-users.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/missingac/documentation/missing-function-ac-03-users.adoc"></div>
 
     <div class="attack-container">
         <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
@@ -92,7 +92,7 @@
 
 <div class="lesson-page-wrapper">
 
-    <div class="adoc-content" th:replace="doc:lessons/missing_ac/documentation/missing-function-ac-04-users-fixed.adoc"></div>
+    <div class="adoc-content" th:replace="doc:lessons/missingac/documentation/missing-function-ac-04-users-fixed.adoc"></div>
 
     <div class="attack-container">
         <div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
