+++ src/test/java/org/owasp/webgoat/lessons/missingac/MissingFunctionACUsersTest.java
@@ -20,42 +20,53 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.missing_ac;
+package org.owasp.webgoat.lessons.missingac;
 
 import org.hamcrest.CoreMatchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.owasp.webgoat.container.plugins.LessonTest;
-import org.owasp.webgoat.lessons.missing_ac.MissingFunctionAC;
+import org.owasp.webgoat.lessons.missingac.MissingFunctionAC;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 
+import static org.hamcrest.Matchers.hasSize;
+import static org.hamcrest.Matchers.is;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-class MissingFunctionYourHashTest extends LessonTest {
+class MissingFunctionACUsersTest extends LessonTest {
 
     @BeforeEach
-    public void setup() {
+    void setup() {
         when(webSession.getCurrentLesson()).thenReturn(new MissingFunctionAC());
         this.mockMvc = MockMvcBuilders.webAppContextSetup(this.wac).build();
     }
 
     @Test
-    void hashDoesNotMatch() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.post("/access-control/user-hash")
-                .param("userHash", "42"))
+    void getUsers() throws Exception {
+        mockMvc.perform(MockMvcRequestBuilders.get("/access-control/users")
+                .header("Content-type", "application/json"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.lessonCompleted", CoreMatchers.is(false)));
+                .andExpect(jsonPath("$[0].username", CoreMatchers.is("Tom")))
+                .andExpect(jsonPath("$[0].userHash", CoreMatchers.is("Mydnhcy00j2b0m6SjmPz6PUxF9WIeO7tzm665GiZWCo=")))
+                .andExpect(jsonPath("$[0].admin", CoreMatchers.is(false)));
     }
 
     @Test
-    void hashMatches() throws Exception {
-        mockMvc.perform(MockMvcRequestBuilders.post("/access-control/user-hash")
-                .param("userHash", "SVtOlaa+ER+w2eoIIVE5/77umvhcsh5V8UyDLUa1Itg="))
+    void addUser() throws Exception {
+        var user = """
+                {"username":"newUser","password":"newUser12","admin": "true"}
+                """;
+        mockMvc.perform(MockMvcRequestBuilders.post("/access-control/users")
+                .header("Content-type", "application/json").content(user))
+                .andExpect(status().isOk());
+
+        mockMvc.perform(MockMvcRequestBuilders.get("/access-control/users")
+                .header("Content-type", "application/json"))
                 .andExpect(status().isOk())
-                .andExpect(jsonPath("$.lessonCompleted", CoreMatchers.is(true)));
+                .andExpect(jsonPath("$.size()", is(4)));
     }
 }
